🔨 refactor: style 분리
