+++ src/pages/LetterReplyPage/index.tsx
@@ -1,11 +1,11 @@
 import { Suspense } from 'react';
 import { useParams } from 'react-router-dom';
-import { css } from '@emotion/react';
 import LoadingSpinner from '@/components/LoadingSpinner';
 import ReplyHeader from './components/ReplyHeader';
 import SentLetter from './sentLetter';
 import ReplyLetter from './replyLetter';
 import BottomButton from './components/BottomButton';
+import style from './styles';
 
 const LetterReplyPage = () => {
   const { letterId } = useParams();
@@ -34,32 +34,3 @@ const LetterReplyPage = () => {
 };
 
 export default LetterReplyPage;
-
-const style = {
-  container: css`
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-  `,
-  content: css`
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100%;
-    padding-inline: 1rem;
-  `,
-  letter: css`
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  `,
-  loadingSpinner: css`
-    display: flex;
-    flex-direction: column;
-    gap: 2rem;
-    justify-content: center;
-    align-items: center;
-    margin-top: 3rem;
-  `,
-};
