+++ src/pages/LetterReplyPage/sentLetter/index.tsx
@@ -1,7 +1,5 @@
-import { css } from '@emotion/react';
 import { CaretRight, CaretLeft } from '@/assets/icons';
 import COLORS from '@/constants/colors';
-import textStyles from '@/styles/textStyles';
 import useBoolean from '@/hooks/useBoolean';
 import Modal from '@/components/Modal';
 import Header from '@/components/Header';
@@ -9,8 +7,9 @@ import LetterCard from '@/components/LetterCard';
 import PolaroidModal from '@/components/PolaroidModal';
 import TagList from '@/pages/LetterReceptionPage/components/TagList';
 import Button from '@/components/Button';
-import useLetterReplyWithTag from '../hooks/useLetterReplyWithTag';
 import LetterContent from '../components/LetterContent';
+import useLetterReplyWithTag from '../hooks/useLetterReplyWithTag';
+import style from './styles';
 interface SentLetterProps {
   letterId: number;
 }
@@ -67,46 +66,3 @@ const SentLetter = ({ letterId }: SentLetterProps) => {
 };
 
 export default SentLetter;
-
-const style = {
-  container: css`
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem 1.25rem;
-    border-radius: 0.5rem;
-    background: #fff;
-    cursor: pointer;
-
-    :hover {
-      background: ${COLORS.gray6};
-    }
-  `,
-  text: css`
-    color: ${COLORS.gray4};
-    ${textStyles.b4m}
-  `,
-  header: css`
-    height: 2.5rem;
-    margin-bottom: 0.5rem;
-    padding-top: 1.25rem;
-    padding-bottom: 0.5rem;
-  `,
-  leftHeader: css`
-    display: flex;
-    gap: 0.625rem;
-    align-items: center;
-  `,
-  icon: css`
-    cursor: pointer;
-  `,
-  headerText: css`
-    color: white;
-    font-weight: 600;
-    font-size: 16px;
-    line-height: 16px;
-  `,
-  content: css`
-    padding-inline: 1rem;
-  `,
-};
