+++ src/pages/LetterReplyPage/sentLetter/styles.ts
@@ -1,0 +1,48 @@
+import { css } from '@emotion/react';
+import COLORS from '@/constants/colors';
+import textStyles from '@/styles/textStyles';
+
+const style = {
+  container: css`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem 1.25rem;
+    border-radius: 0.5rem;
+    background: #fff;
+    cursor: pointer;
+
+    :hover {
+      background: ${COLORS.gray6};
+    }
+  `,
+  text: css`
+    color: ${COLORS.gray4};
+    ${textStyles.b4m}
+  `,
+  header: css`
+    height: 2.5rem;
+    margin-bottom: 0.5rem;
+    padding-top: 1.25rem;
+    padding-bottom: 0.5rem;
+  `,
+  leftHeader: css`
+    display: flex;
+    gap: 0.625rem;
+    align-items: center;
+  `,
+  icon: css`
+    cursor: pointer;
+  `,
+  headerText: css`
+    color: white;
+    font-weight: 600;
+    font-size: 16px;
+    line-height: 16px;
+  `,
+  content: css`
+    padding-inline: 1rem;
+  `,
+};
+
+export default style;
